✨ Implement BEMA algorithm for reference model updates during DPO training
